drawCircle and getColor added
drawCircle allows the caller to draw a circle of the color that they
specify as well as whether the circle should be filled.

getColor retrieves the color that should be used based on the color that
is requested and the index of the current pixel to draw. This matches
what drawLine uses where even indeces are shown as WHITE for the color
GRAY

A basic example:
display.drawCircle(20, 25, 5, WHITE, true);
display.update();